Add deconvolved kernel regression compute_state path
