+++ recovar/heterogeneity/adaptive_kernel_discretization.py
@@ -13,718 +13,15 @@ import equinox as eqx
 import jax
 import jax.numpy as jnp
 import numpy as np
-from scipy import special
 
 import recovar.core.forward as core_forward
 import recovar.core.fourier_transform_utils as fourier_transform_utils
 from recovar import core, jax_config, utils
 from recovar.core import linalg
 from recovar.core.configs import ForwardModelConfig
-from recovar.cuda_backproject import custom_cuda_requested
-from recovar.data_io import cryoem_dataset
-from recovar.reconstruction import noise, regularization, relion_functions
+from recovar.reconstruction import noise, regularization
 
 logger = logging.getLogger(__name__)
-
-_SENTINEL = object()
-DEFAULT_DECONV_LAMBDA_GRID = np.asarray(
-    [1.0, 1.25, 1.5, 2.0, 3.0, 4.0, 6.0, 8.0, 12.0, 16.0],
-    dtype=np.float32,
-)
-_DECONV_EPS = 1e-6
-
-
-def epanechnikov_kernel_1d(u):
-    """Evaluate the one-dimensional Epanechnikov kernel."""
-    u = np.asarray(u)
-    return np.where(np.abs(u) < 1, 0.75 * (1 - u**2), 0)
-
-
-def epanechnikov_deconvolution_kernel_1d(u, lambda_val):
-    """Gaussian-error Epanechnikov deconvolution kernel for 1D latent variables."""
-    u = np.asarray(u, dtype=np.float64)
-    lambda_val = np.asarray(lambda_val, dtype=np.float64)
-    lambda_val = np.maximum(lambda_val, _DECONV_EPS)
-    abs_u = np.abs(u)
-    z = (1j * lambda_val * abs_u - 1.0 / lambda_val) / np.sqrt(2.0)
-    with np.errstate(over="ignore", invalid="ignore"):
-        weights = (
-            -lambda_val
-            / np.sqrt(2.0 * np.pi)
-            * np.exp(1.0 / (2.0 * lambda_val**2))
-            * np.imag(np.exp(-1j * abs_u) * special.wofz(z))
-        )
-    return weights
-
-
-def deconvolution_weights_1d(latent_diff, latent_precision, h):
-    """Return signed deconvolution weights for one target/bandwidth.
-
-    ``latent_precision`` is the RECOVAR latent precision.  In 1D the latent
-    noise variance is its reciprocal; do not invert it before calling this.
-    Invalid precision or non-finite weights are set to zero.
-    """
-    return deconvolution_weights_1d_many(latent_diff, latent_precision, np.asarray([h], dtype=np.float64))[0]
-
-
-def deconvolution_weights_1d_many(latent_diff, latent_precision, h_grid):
-    """Return signed deconvolution weights for several bandwidths at once.
-
-    Output shape is ``(len(h_grid), n_images)``.
-    """
-    latent_diff = np.asarray(latent_diff, dtype=np.float64).reshape(-1)
-    latent_precision = np.asarray(latent_precision, dtype=np.float64).reshape(-1)
-    if latent_diff.shape != latent_precision.shape:
-        raise ValueError(
-            "latent_diff and latent_precision must have the same flattened shape, "
-            f"got {latent_diff.shape} and {latent_precision.shape}"
-        )
-    h_grid = np.asarray(h_grid, dtype=np.float64).reshape(-1)
-    if h_grid.size == 0 or not np.all(np.isfinite(h_grid)) or np.any(h_grid <= 0):
-        raise ValueError(f"h_grid must contain finite positive values, got {h_grid}")
-
-    valid_precision = np.isfinite(latent_precision) & (latent_precision > 0)
-    latent_noise_variance = np.zeros_like(latent_precision, dtype=np.float64)
-    latent_noise_variance[valid_precision] = 1.0 / latent_precision[valid_precision]
-    latent_noise_std = np.full_like(latent_precision, np.inf, dtype=np.float64)
-    latent_noise_std[valid_precision] = np.sqrt(latent_noise_variance[valid_precision])
-
-    weights = np.zeros((h_grid.size, latent_diff.size), dtype=np.float64)
-    valid_inputs = valid_precision & np.isfinite(latent_diff)
-    if np.any(valid_inputs):
-        u = latent_diff[valid_inputs][None, :] / h_grid[:, None]
-        lambda_i = h_grid[:, None] / np.maximum(latent_noise_std[valid_inputs][None, :], _DECONV_EPS)
-        weights[:, valid_inputs] = epanechnikov_deconvolution_kernel_1d(u, lambda_i)
-    valid_weights = valid_inputs[None, :] & np.isfinite(weights)
-    weights = np.where(valid_weights, weights, 0.0)
-    with np.errstate(over="ignore", invalid="ignore"):
-        weights32 = weights.astype(np.float32)
-    valid_weights32 = np.isfinite(weights32)
-    if not np.all(valid_weights32):
-        weights32 = np.where(valid_weights32, weights32, 0.0)
-        valid_weights = valid_weights & valid_weights32
-    if not np.all(valid_weights):
-        logger.warning(
-            "Zeroing %d invalid deconvolution weights out of %d",
-            int(weights.size - np.count_nonzero(valid_weights)),
-            int(weights.size),
-        )
-    return weights32
-
-
-def _coerce_deconv_lambda_grid(lambda_grid):
-    if lambda_grid is None:
-        return DEFAULT_DECONV_LAMBDA_GRID.copy()
-    arr = np.asarray(lambda_grid, dtype=np.float32).reshape(-1)
-    if arr.size == 0 or not np.all(np.isfinite(arr)) or np.any(arr <= 0):
-        raise ValueError(f"lambda_grid must contain finite positive values, got {lambda_grid}")
-    return arr
-
-
-def deconvolution_bandwidths_1d(latent_precision, lambda_grid=None, sigma_ref=None):
-    """Return ``(lambda_grid, h_grid, sigma_ref)`` for finite positive precision."""
-    lambda_grid = _coerce_deconv_lambda_grid(lambda_grid)
-    if sigma_ref is None:
-        latent_precision = np.asarray(latent_precision, dtype=np.float64).reshape(-1)
-        valid_precision = np.isfinite(latent_precision) & (latent_precision > 0)
-        if not np.any(valid_precision):
-            raise ValueError("No finite positive latent precision values for deconvolution bandwidth selection.")
-        latent_noise_variance = 1.0 / latent_precision[valid_precision]
-        latent_noise_std = np.sqrt(latent_noise_variance)
-        sigma_ref = float(np.median(latent_noise_std))
-    else:
-        sigma_ref = float(sigma_ref)
-    if not np.isfinite(sigma_ref) or sigma_ref <= 0:
-        raise ValueError(f"Invalid deconvolution sigma_ref={sigma_ref}")
-    return lambda_grid, (lambda_grid.astype(np.float64) * sigma_ref), sigma_ref
-
-
-def _auto_deconv_lambda_batch_size(n_lambdas, half_volume_size, complex_dtype, real_dtype):
-    """Pick how many lambda candidates to backproject together."""
-    per_lambda_gb = (
-        half_volume_size * (np.dtype(complex_dtype).itemsize + np.dtype(real_dtype).itemsize) / (1024**3)
-    )
-    if per_lambda_gb <= 0:
-        return 1
-    target_gb = max(1.0, 0.20 * float(utils.get_gpu_memory_total()))
-    return max(1, min(int(n_lambdas), 64, int(target_gb / per_lambda_gb)))
-
-
-def _auto_deconv_per_image_batch_size(batch_size, n_weight_sets, half_volume_size, complex_dtype, real_dtype):
-    """Cap image batch size for dense per-image backprojection temporaries."""
-    if not (custom_cuda_requested() and jax.default_backend() == "gpu"):
-        return batch_size
-
-    bytes_per_image = half_volume_size * (np.dtype(complex_dtype).itemsize + np.dtype(real_dtype).itemsize)
-    bytes_per_weight_set = bytes_per_image
-    target_bytes = utils.get_gpu_memory_total() * (1024**3) * 0.45
-    target_bytes -= 2 * n_weight_sets * bytes_per_weight_set
-    memory_limited = max(1, int(target_bytes // max(1, bytes_per_image)))
-    return max(1, min(int(batch_size), int(memory_limited), 256))
-
-
-def _coerce_1d_latent_differences(latent_differences):
-    arr = np.asarray(latent_differences, dtype=np.float32)
-    if arr.ndim == 1:
-        return arr
-    if arr.ndim == 2 and arr.shape[1] == 1:
-        return arr[:, 0]
-    raise NotImplementedError(f"Deconvolved kernel regression only supports zdim=1; got shape {arr.shape}")
-
-
-def _coerce_1d_latent_precision(latent_precision):
-    arr = np.asarray(latent_precision, dtype=np.float32)
-    if arr.ndim == 1:
-        return arr
-    if arr.ndim == 2 and arr.shape[1] == 1:
-        return arr[:, 0]
-    if arr.ndim == 3 and arr.shape[1:] == (1, 1):
-        return arr[:, 0, 0]
-    raise NotImplementedError(f"Deconvolved kernel regression only supports zdim=1 precision; got shape {arr.shape}")
-
-
-def _expand_tilt_latent_array_to_images(experiment_dataset, values, name):
-    values = np.asarray(values)
-    if values.shape[0] == experiment_dataset.n_images:
-        return values
-    if (
-        getattr(experiment_dataset, "tilt_series_flag", False)
-        and hasattr(experiment_dataset, "tilt_particles")
-        and values.shape[0] != experiment_dataset.n_images
-    ):
-        per_image = np.empty(experiment_dataset.n_images, dtype=values.dtype)
-        for p_idx, tilt_inds in enumerate(experiment_dataset.tilt_particles):
-            per_image[tilt_inds] = values[p_idx]
-        return per_image
-    raise ValueError(f"{name} length {values.shape[0]} does not match dataset n_images={experiment_dataset.n_images}")
-
-
-def _effective_heterogeneity_memory_budget(avail_gb):
-    """Apply the fallback-path safety margin to the batch-size budget."""
-    if custom_cuda_requested():
-        return avail_gb
-
-    scaled_gb = max(1.0, avail_gb / 3.0)
-    logger.info(
-        "RECOVAR_DISABLE_CUDA is active - scaling heterogeneity-kernel "
-        "memory budget to 1/3 of available (%.1f GB) to account for the "
-        "JAX-native fallback path's higher per-image memory cost.",
-        scaled_gb,
-    )
-    return scaled_gb
-
-
-def _pad_noise_variance_for_fixed_batch(noise_variance, current_batch_size, target_batch_size):
-    """Pad per-image noise rows so padded batch items contribute zero weight."""
-    if noise_variance is None:
-        return None
-
-    arr = np.asarray(noise_variance)
-    if arr.ndim == 1:
-        arr = arr.reshape(1, -1)
-    if arr.ndim != 2:
-        raise ValueError(f"noise_variance must be a 1D or 2D array when padding batches, got shape {arr.shape}")
-
-    if arr.shape[0] == 1:
-        arr = np.repeat(arr, current_batch_size, axis=0)
-    elif arr.shape[0] != current_batch_size:
-        raise ValueError(
-            f"noise_variance rows must match batch size when padding batches: {arr.shape[0]} != {current_batch_size}"
-        )
-
-    if current_batch_size == target_batch_size:
-        return arr
-
-    padded = np.full((target_batch_size, arr.shape[1]), np.inf, dtype=arr.dtype)
-    padded[:current_batch_size] = arr
-    return padded
-
-
-def _pad_heterogeneity_kernel_batch(
-    images,
-    rotation_matrices,
-    translations,
-    ctf_params,
-    noise_variance,
-    *,
-    target_batch_size,
-):
-    """Pad the final image batch to a fixed size to avoid JIT recompilation.
-
-    The heterogeneity kernel accumulators are sums, so padded rows can be made
-    exactly neutral by assigning them infinite noise variance.
-
-    JAX device arrays stay on device; numpy arrays stay on the host.
-    """
-    current_batch_size = int(images.shape[0])
-    if current_batch_size == target_batch_size:
-        return images, rotation_matrices, translations, ctf_params, noise_variance
-    if current_batch_size > target_batch_size:
-        raise ValueError(f"batch size {current_batch_size} exceeds target_batch_size {target_batch_size}")
-
-    pad = target_batch_size - current_batch_size
-
-    def _pad_zeros(arr, n):
-        if isinstance(arr, jax.Array):
-            return jnp.concatenate([arr, jnp.zeros((n, *arr.shape[1:]), dtype=arr.dtype)])
-        arr = np.asarray(arr)
-        return np.concatenate([arr, np.zeros((n, *arr.shape[1:]), dtype=arr.dtype)])
-
-    def _pad_repeat_first(arr, n):
-        if isinstance(arr, jax.Array):
-            return jnp.concatenate([arr, jnp.repeat(arr[:1], n, axis=0)])
-        arr = np.asarray(arr)
-        return np.concatenate([arr, np.repeat(arr[:1], n, axis=0)])
-
-    images = _pad_zeros(images, pad)
-    rotation_matrices = _pad_repeat_first(rotation_matrices, pad)
-    translations = _pad_repeat_first(translations, pad)
-    ctf_params = _pad_repeat_first(ctf_params, pad)
-    noise_variance = _pad_noise_variance_for_fixed_batch(
-        noise_variance,
-        current_batch_size,
-        target_batch_size,
-    )
-    return images, rotation_matrices, translations, ctf_params, noise_variance
-
-
-def _pad_image_weights_for_fixed_batch(image_weights, current_batch_size, target_batch_size):
-    """Pad image weights with zeros so synthetic padded rows are neutral."""
-    arr = np.asarray(image_weights)
-    if arr.ndim != 1:
-        arr = arr.reshape(-1)
-    if arr.shape[0] != current_batch_size:
-        raise ValueError(f"image_weights length must match batch size: {arr.shape[0]} != {current_batch_size}")
-    if current_batch_size == target_batch_size:
-        return arr
-    if current_batch_size > target_batch_size:
-        raise ValueError(f"batch size {current_batch_size} exceeds target_batch_size {target_batch_size}")
-    padded = np.zeros((target_batch_size,), dtype=arr.dtype)
-    padded[:current_batch_size] = arr
-    return padded
-
-
-def _pad_image_weight_matrix_for_fixed_batch(image_weights, current_batch_size, target_batch_size):
-    """Pad a ``(n_weights, n_images)`` weight matrix with zero image rows."""
-    arr = np.asarray(image_weights)
-    if arr.ndim != 2:
-        raise ValueError(f"image_weights must have shape (n_weights, n_images), got {arr.shape}")
-    if arr.shape[1] != current_batch_size:
-        raise ValueError(f"image_weights image axis must match batch size: {arr.shape[1]} != {current_batch_size}")
-    if current_batch_size == target_batch_size:
-        return arr
-    if current_batch_size > target_batch_size:
-        raise ValueError(f"batch size {current_batch_size} exceeds target_batch_size {target_batch_size}")
-    padded = np.zeros((arr.shape[0], target_batch_size), dtype=arr.dtype)
-    padded[:, :current_batch_size] = arr
-    return padded
-
-
-def _iter_processed_half_batches(experiment_dataset, raw_batches):
-    """Prefetch raw batches and overlap half-spectrum preprocessing with GPU work.
-
-    Keeping preprocessing out of the generic dataset iterator keeps the data IO
-    API clean, but the heterogeneity kernel still needs the old overlap between
-    CPU-side FFT preprocessing and GPU backprojection.
-    """
-    from concurrent.futures import ThreadPoolExecutor
-
-    raw_iter = iter(raw_batches)
-
-    def _prepare_next():
-        batch = next(raw_iter, _SENTINEL)
-        if batch is _SENTINEL:
-            return _SENTINEL
-        images, rotation_matrices, translations, ctf_params, noise_variance, particle_indices, image_indices = batch
-        half_images = experiment_dataset.process_images_half(images)
-        return (
-            half_images,
-            rotation_matrices,
-            translations,
-            ctf_params,
-            noise_variance,
-            particle_indices,
-            image_indices,
-        )
-
-    with ThreadPoolExecutor(max_workers=1) as pool:
-        future = pool.submit(_prepare_next)
-        while True:
-            result = future.result()
-            if result is _SENTINEL:
-                return
-            future = pool.submit(_prepare_next)
-            yield result
-
-
-def _process_images_half_fast(images, dataset):
-    """Convert raw spatial images to half-spectrum using JIT'd rfft2.
-
-    Unlike ``dataset.process_images_half`` which does a full complex FFT
-    then discards half, this uses ``padded_rfft`` (rfft2) directly —
-    half the work, JIT-fused into a single GPU kernel.  The tiny numerical
-    difference vs the full-FFT path is irrelevant for volume reconstruction.
-
-    Falls back to ``process_images_half`` for non-standard image sources
-    (e.g. test fixtures) that lack ``data_multiplier`` / ``padding``.
-    """
-    src = getattr(dataset, "image_source", None)
-    if src is not None and hasattr(src, "data_multiplier") and hasattr(src, "padding"):
-        from recovar.core import padding as pad
-
-        images = jnp.asarray(images) * src.data_multiplier
-        return pad.padded_rfft(images, dataset.grid_size, src.padding)
-    return dataset.process_images_half(images)
-
-
-def _heterogeneity_kernel_batch_from_fft(
-    config: ForwardModelConfig,
-    images,
-    ctf_params,
-    rotation_matrices,
-    translations,
-    noise_variance,
-    Ft_y: jax.Array = None,
-    Ft_ctf: jax.Array = None,
-    upsample_ctf: bool = True,
-):
-    """Backproject half-spectrum images into heterogeneity kernel accumulators.
-
-    Expects ``images`` in half-image (rfft) format. Accumulates into
-    half-volume layout for memory efficiency.
-
-    Parameters
-    ----------
-    upsample_ctf : bool
-        If True, evaluate CTF² on a 2x-upsampled grid and box-filter back to
-        native resolution before backprojecting.  Reduces aliasing in the CTF
-        weight accumulator.  Default True (matches legacy behaviour).
-    """
-    return _heterogeneity_kernel_batch_from_fft_explicit(
-        config,
-        images,
-        ctf_params,
-        rotation_matrices,
-        translations,
-        noise_variance,
-        Ft_y=Ft_y,
-        Ft_ctf=Ft_ctf,
-        upsample_ctf=upsample_ctf,
-    )
-
-
-def _heterogeneity_weighted_kernel_batch_from_fft(
-    config: ForwardModelConfig,
-    images,
-    ctf_params,
-    rotation_matrices,
-    translations,
-    noise_variance,
-    image_weights,
-    Ft_y: jax.Array = None,
-    Ft_ctf: jax.Array = None,
-    upsample_ctf: bool = True,
-):
-    """Backproject half-spectrum images with one scalar signed weight per image."""
-    return _heterogeneity_weighted_kernel_batch_from_fft_explicit(
-        config,
-        images,
-        ctf_params,
-        rotation_matrices,
-        translations,
-        noise_variance,
-        image_weights,
-        Ft_y=Ft_y,
-        Ft_ctf=Ft_ctf,
-        upsample_ctf=upsample_ctf,
-    )
-
-
-def _can_use_cuda_batch_backproject(config: ForwardModelConfig) -> bool:
-    return custom_cuda_requested() and jax.default_backend() == "gpu" and core.decide_order(config.disc_type) <= 1
-
-
-def _heterogeneity_weighted_kernel_batch_many_from_fft(
-    config: ForwardModelConfig,
-    images,
-    ctf_params,
-    rotation_matrices,
-    translations,
-    noise_variance,
-    image_weights,
-    Ft_y: jax.Array = None,
-    Ft_ctf: jax.Array = None,
-    upsample_ctf: bool = True,
-):
-    """Backproject one image batch into several weighted accumulators.
-
-    ``image_weights`` has shape ``(n_weight_sets, n_images)``.  On GPU with
-    RECOVAR's CUDA kernel available this uses ``batch_backproject`` so all
-    weight sets share the translated images, CTFs, rotations, and interpolation
-    coordinates.  The fallback preserves correctness by looping over weight
-    sets through the scalar helper.
-    """
-    if _can_use_cuda_batch_backproject(config):
-        return _heterogeneity_weighted_kernel_batch_many_from_fft_explicit(
-            config,
-            images,
-            ctf_params,
-            rotation_matrices,
-            translations,
-            noise_variance,
-            image_weights,
-            Ft_y=Ft_y,
-            Ft_ctf=Ft_ctf,
-            upsample_ctf=upsample_ctf,
-        )
-
-    image_weights = np.asarray(image_weights)
-    y_rows = []
-    ctf_rows = []
-    for weight_idx in range(image_weights.shape[0]):
-        y_row, ctf_row = _heterogeneity_weighted_kernel_batch_from_fft(
-            config,
-            images,
-            ctf_params,
-            rotation_matrices,
-            translations,
-            noise_variance,
-            image_weights[weight_idx],
-            Ft_y=None if Ft_y is None else Ft_y[weight_idx],
-            Ft_ctf=None if Ft_ctf is None else Ft_ctf[weight_idx],
-            upsample_ctf=upsample_ctf,
-        )
-        y_rows.append(y_row)
-        ctf_rows.append(ctf_row)
-    return jnp.stack(y_rows, axis=0), jnp.stack(ctf_rows, axis=0)
-
-
-@eqx.filter_jit
-def _heterogeneity_kernel_batch_from_fft_explicit(
-    config: ForwardModelConfig,
-    images,
-    ctf_params,
-    rotation_matrices,
-    translations,
-    noise_variance,
-    Ft_y: jax.Array = None,
-    Ft_ctf: jax.Array = None,
-    upsample_ctf: bool = False,
-):
-    from recovar.core.geometry import translate_images
-    from recovar.reconstruction import noise as noise_mod
-
-    half_images = translate_images(images, translations, config.image_shape, half_image=True)
-    noise_half = noise_mod.to_batched_half_pixel_noise(
-        noise_variance, config.image_shape, batch_size=half_images.shape[0]
-    )
-    half_images = half_images / noise_half
-
-    # Pre-compute CTF once (reused for both image and weight backprojections)
-    ctf = config.compute_ctf_half(ctf_params)
-    if not config.premultiplied_ctf:
-        half_images = half_images * ctf
-
-    # TODO: remove max_r=None once the default max_r clipping is resolved globally.
-    # Currently needed because the old code had no sphere clipping and the default
-    # max_r=image_shape[0]//2-1 discards the outermost frequency shell.
-    Ft_y = core.adjoint_slice_volume(
-        half_images,
-        rotation_matrices,
-        config.image_shape,
-        config.volume_shape,
-        config.disc_type,
-        volume=Ft_y,
-        half_image=True,
-        half_volume=True,
-        max_r=None,
-    )
-
-    if upsample_ctf:
-        from recovar.core.ctf import compute_antialiased_ctf_squared
-
-        ctf_half = (
-            compute_antialiased_ctf_squared(
-                config.ctf,
-                ctf_params,
-                config.image_shape,
-                config.voxel_size,
-                half_image=True,
-            )
-            / noise_half
-        )
-    else:
-        ctf_half = ctf**2 / noise_half
-
-    Ft_ctf = core.adjoint_slice_volume(
-        ctf_half,
-        rotation_matrices,
-        config.image_shape,
-        config.volume_shape,
-        config.disc_type,
-        volume=Ft_ctf,
-        half_image=True,
-        half_volume=True,
-        max_r=None,  # TODO: see above
-    )
-    return Ft_y, Ft_ctf.real
-
-
-@eqx.filter_jit
-def _heterogeneity_weighted_kernel_batch_many_from_fft_explicit(
-    config: ForwardModelConfig,
-    images,
-    ctf_params,
-    rotation_matrices,
-    translations,
-    noise_variance,
-    image_weights,
-    Ft_y: jax.Array = None,
-    Ft_ctf: jax.Array = None,
-    upsample_ctf: bool = False,
-):
-    from recovar.core.geometry import translate_images
-    from recovar.reconstruction import noise as noise_mod
-    from recovar.cuda_backproject import per_image_backproject
-
-    half_images = translate_images(images, translations, config.image_shape, half_image=True)
-    n_images = half_images.shape[0]
-    half_images = half_images.reshape(n_images, -1)
-    weights = jnp.asarray(image_weights, dtype=half_images.real.dtype)
-
-    noise_half = noise_mod.to_batched_half_pixel_noise(
-        noise_variance, config.image_shape, batch_size=half_images.shape[0]
-    ).reshape(n_images, -1)
-    half_images = half_images / noise_half
-
-    ctf = config.compute_ctf_half(ctf_params).reshape(n_images, -1)
-    if not config.premultiplied_ctf:
-        half_images = half_images * ctf
-
-    if Ft_y is None:
-        vol_shape = fourier_transform_utils.volume_shape_to_half_volume_shape(config.volume_shape)
-        Ft_y = jnp.zeros((weights.shape[0], int(np.prod(vol_shape))), dtype=half_images.dtype)
-
-    per_image_y = jnp.zeros((n_images, Ft_y.shape[1]), dtype=half_images.dtype)
-    per_image_y = per_image_backproject(
-        per_image_y,
-        half_images,
-        rotation_matrices,
-        config.image_shape,
-        config.volume_shape,
-        order=core.decide_order(config.disc_type),
-        half_image=True,
-        half_volume=True,
-        max_r=None,
-    )
-    Ft_y = Ft_y + weights.astype(per_image_y.real.dtype) @ per_image_y
-
-    if upsample_ctf:
-        from recovar.core.ctf import compute_antialiased_ctf_squared
-
-        ctf_half = (
-            compute_antialiased_ctf_squared(
-                config.ctf,
-                ctf_params,
-                config.image_shape,
-                config.voxel_size,
-                half_image=True,
-            ).reshape(n_images, -1)
-            / noise_half
-        )
-    else:
-        ctf_half = ctf**2 / noise_half
-
-    if Ft_ctf is None:
-        vol_shape = fourier_transform_utils.volume_shape_to_half_volume_shape(config.volume_shape)
-        Ft_ctf = jnp.zeros((weights.shape[0], int(np.prod(vol_shape))), dtype=ctf_half.real.dtype)
-
-    per_image_ctf = jnp.zeros((n_images, Ft_ctf.shape[1]), dtype=Ft_ctf.dtype)
-    per_image_ctf = per_image_backproject(
-        per_image_ctf,
-        ctf_half.astype(Ft_ctf.dtype),
-        rotation_matrices,
-        config.image_shape,
-        config.volume_shape,
-        order=core.decide_order(config.disc_type),
-        half_image=True,
-        half_volume=True,
-        max_r=None,
-    )
-    Ft_ctf = Ft_ctf + weights.astype(Ft_ctf.dtype) @ per_image_ctf
-    return Ft_y, Ft_ctf.real
-
-
-@eqx.filter_jit
-def _heterogeneity_weighted_kernel_batch_from_fft_explicit(
-    config: ForwardModelConfig,
-    images,
-    ctf_params,
-    rotation_matrices,
-    translations,
-    noise_variance,
-    image_weights,
-    Ft_y: jax.Array = None,
-    Ft_ctf: jax.Array = None,
-    upsample_ctf: bool = False,
-):
-    from recovar.core.geometry import translate_images
-    from recovar.reconstruction import noise as noise_mod
-
-    half_images = translate_images(images, translations, config.image_shape, half_image=True)
-    weights = jnp.asarray(image_weights, dtype=half_images.real.dtype)
-    weights = jnp.reshape(weights, (weights.shape[0],) + (1,) * (half_images.ndim - 1))
-    noise_half = noise_mod.to_batched_half_pixel_noise(
-        noise_variance, config.image_shape, batch_size=half_images.shape[0]
-    )
-    half_images = half_images / noise_half
-
-    ctf = config.compute_ctf_half(ctf_params)
-    if not config.premultiplied_ctf:
-        half_images = half_images * ctf
-    half_images = half_images * weights
-
-    Ft_y = core.adjoint_slice_volume(
-        half_images,
-        rotation_matrices,
-        config.image_shape,
-        config.volume_shape,
-        config.disc_type,
-        volume=Ft_y,
-        half_image=True,
-        half_volume=True,
-        max_r=None,
-    )
-
-    if upsample_ctf:
-        from recovar.core.ctf import compute_antialiased_ctf_squared
-
-        ctf_half = (
-            compute_antialiased_ctf_squared(
-                config.ctf,
-                ctf_params,
-                config.image_shape,
-                config.voxel_size,
-                half_image=True,
-            )
-            / noise_half
-        )
-    else:
-        ctf_half = ctf**2 / noise_half
-    ctf_half = ctf_half * weights
-
-    Ft_ctf = core.adjoint_slice_volume(
-        ctf_half,
-        rotation_matrices,
-        config.image_shape,
-        config.volume_shape,
-        config.disc_type,
-        volume=Ft_ctf,
-        half_image=True,
-        half_volume=True,
-        max_r=None,
-    )
-    return Ft_y, Ft_ctf.real
 
 
 def make_X_mat(rotation_matrices, volume_shape, image_shape, pol_degree=0, dtype=np.float32):
@@ -2079,504 +1376,6 @@ def heterogeneous_reconstruction_fixed_variance(
         )  # XWX, Z, F, alpha
     else:
         return np.asarray(weights_opt), np.asarray(disc_choices), np.asarray(residuals_averaged)
-
-
-def naive_heterogeneity_scheme_relion_style(
-    experiment_dataset,
-    noise_variance,
-    signal_variance,
-    heterogeneity_distances,
-    heterogeneity_bins,
-    batch_size=100,
-    tau=None,
-    compute_lhs_rhs=False,
-    grid_correct=True,
-    disc_type="linear_interp",
-):
-    import gc
-
-    estimates = []
-    lhs, rhs = [], []
-    og_contrast = experiment_dataset.get_ctf_column(8)
-    idx = 0
-    for residual_threshold in heterogeneity_bins:
-        from recovar.data_io import cryoem_dataset
-
-        good_indices = heterogeneity_distances <= residual_threshold
-        # utils.report_memory_device(logger=logger)
-
-        test_dataset = cryoem_dataset.subsample_cryoem_dataset(experiment_dataset, good_indices)
-        utils.report_memory_device(logger=logger)
-        from recovar.reconstruction import relion_functions
-
-        if compute_lhs_rhs:
-            Ft_ctf, F_ty = relion_functions.relion_style_triangular_kernel(
-                test_dataset, noise_variance.astype(np.float32), batch_size, disc_type=disc_type, upsampling_factor=1
-            )
-
-            kernel_type = "triangular" if disc_type == "linear_interp" else "square"
-            estimate = relion_functions.post_process_from_filter_v2(
-                Ft_ctf,
-                F_ty,
-                test_dataset.volume_shape,
-                1,
-                tau=tau,
-                kernel=kernel_type,
-                use_spherical_mask=False,
-                grid_correct=grid_correct,
-                gridding_correct="square",
-                kernel_width=1,
-            )
-
-            lhs.append(np.array(Ft_ctf))
-            rhs.append(np.array(F_ty))
-        else:
-            estimate = relion_functions.relion_reconstruct(
-                test_dataset,
-                noise_variance,
-                batch_size,
-                "linear_interp",
-                use_spherical_mask=True,
-                upsampling_factor=2,
-                grid_correct=grid_correct,
-                gridding_correct="square",
-                tau=tau,
-            )
-
-        test_dataset.delete()
-        del test_dataset
-        gc.collect()
-
-        logger.info("Number of images used in estimator %d: %s", idx, np.sum(good_indices))
-        idx += 1
-        estimates.append(np.array(estimate.reshape(-1)))
-
-    estimates = np.array(estimates)
-    if compute_lhs_rhs:
-        return estimates, lhs, rhs
-
-    return estimates
-
-
-def less_naive_heterogeneity_scheme_relion_style(
-    experiment_dataset,
-    noise_variance,
-    signal_variance,
-    heterogeneity_distances,
-    heterogeneity_bins,
-    batch_size=100,
-    tau=None,
-    compute_lhs_rhs=False,
-    grid_correct=True,
-    disc_type="linear_interp",
-):
-
-    estimates = []
-    XWX, F = precompute_triangular_kernel(
-        experiment_dataset,
-        noise_variance.astype(np.float32),
-        pol_degree=0,
-        heterogeneity_distances=heterogeneity_distances,
-        heterogeneity_bins=heterogeneity_bins,
-    )
-
-    XWX = XWX[:, 0, :].T
-    F = F[:, 0, :].T
-
-    kernel_type = "triangular" if disc_type == "linear_interp" else "square"
-    for idx in range(heterogeneity_bins.size):
-        from recovar.reconstruction import relion_functions
-
-        estimate = relion_functions.post_process_from_filter_v2(
-            XWX[idx],
-            F[idx],
-            experiment_dataset.volume_shape,
-            2,
-            tau=tau,
-            kernel=kernel_type,
-            use_spherical_mask=True,
-            grid_correct=grid_correct,
-            gridding_correct="square",
-            kernel_width=1,
-            input_half_volume=True,
-        )
-        estimates.append(np.array(estimate.reshape(-1)))
-    estimates = np.array(estimates)
-
-    return estimates
-
-
-def even_less_naive_heterogeneity_scheme_relion_style(
-    experiment_dataset,
-    signal_variance,
-    heterogeneity_distances,
-    heterogeneity_bins,
-    batch_size=None,
-    tau=None,
-    compute_lhs_rhs=False,
-    grid_correct=True,
-    disc_type="linear_interp",
-    use_spherical_mask=True,
-    return_lhs_rhs=False,
-    heterogeneity_kernel="parabola",
-    upsampling_factor=None,
-    return_real_space=False,
-    use_fast_rfft=False,
-):
-    bins = heterogeneity_bins
-
-    # For tilt-series datasets, distances may be per-particle while the kernel
-    # iterates per-image.  Expand to per-image so bin assignments match image indices.
-    if (
-        getattr(experiment_dataset, "tilt_series_flag", False)
-        and hasattr(experiment_dataset, "tilt_particles")
-        and heterogeneity_distances.shape[0] != experiment_dataset.n_images
-    ):
-        per_image = np.empty(experiment_dataset.n_images, dtype=heterogeneity_distances.dtype)
-        for p_idx, tilt_inds in enumerate(experiment_dataset.tilt_particles):
-            per_image[tilt_inds] = heterogeneity_distances[p_idx]
-        heterogeneity_distances = per_image
-
-    inds = np.digitize(heterogeneity_distances, bins, right=True).astype(np.int32)
-    n_bins = bins.size
-
-    if upsampling_factor is not None:
-        upsampled_vol_shape = tuple(3 * [experiment_dataset.grid_size * upsampling_factor])
-    else:
-        upsampled_vol_shape = tuple((experiment_dataset.grid_size,) * 3)
-    half_volume_size = np.prod(volume_shape_to_half_volume_shape(upsampled_vol_shape))
-
-    # Accumulate on CPU to avoid OOM from JAX immutable-array copies at large
-    # grid sizes (256^3 × 50 bins ≈ 5 GB; .at[].set() would double that).
-    # Transferred to GPU only for the final matmul below.
-    rhs_all = np.zeros((n_bins, half_volume_size), dtype=experiment_dataset.dtype)
-    lhs_all = np.zeros((n_bins, half_volume_size), dtype=experiment_dataset.dtype_real)
-
-    # Auto-compute batch size based on GPU memory if not specified.
-    #
-    # The default formula is calibrated for the custom CUDA backproject kernel,
-    # which scatters in place. The JAX-native fallback path materializes
-    # per-image volumes for the whole batch before reducing, so the working set
-    # is materially larger for the same batch size.
-    if batch_size is None:
-        accum_gb = utils.get_size_in_gb(rhs_all) + utils.get_size_in_gb(lhs_all)
-        avail_gb = _effective_heterogeneity_memory_budget(max(1.0, utils.get_gpu_memory_total() - accum_gb))
-        batch_size = int(utils.get_image_batch_size(experiment_dataset.grid_size, avail_gb))
-    logger.info("batch size in heterogeneity kernel: %s", batch_size)
-
-    if upsampling_factor is not None:
-        config = ForwardModelConfig.from_dataset(
-            experiment_dataset, disc_type=disc_type, upsampling_factor=upsampling_factor
-        )
-    else:
-        config = ForwardModelConfig.from_dataset(experiment_dataset, disc_type=disc_type, upsampling_factor=2)
-
-    image_inds_by_bin = [np.flatnonzero(inds == bin_idx).astype(np.int32) for bin_idx in range(n_bins)]
-    # Pre-allocate accumulators once (reused across bins)
-    Ft_y_acc = jnp.zeros(half_volume_size, dtype=experiment_dataset.dtype)
-    Ft_ctf_acc = jnp.zeros(half_volume_size, dtype=experiment_dataset.dtype_real)
-    for bin_idx, image_inds in enumerate(image_inds_by_bin):
-        if image_inds.size == 0:
-            continue
-
-        Ft_y_acc = jnp.zeros_like(Ft_y_acc)
-        Ft_ctf_acc = jnp.zeros_like(Ft_ctf_acc)
-        raw_batches = experiment_dataset.iter_batches(
-            batch_size,
-            noise_model=experiment_dataset.noise,
-            noise_half=False,
-            indices=image_inds,
-        )
-        for (
-            raw_images,
-            rotation_matrices,
-            translations,
-            ctf_params,
-            noise_variance,
-            _particle_indices,
-            _image_indices,
-        ) in raw_batches:
-            if use_fast_rfft:
-                # Pad BEFORE rfft so padded_rfft always sees a fixed batch
-                # shape — avoids JIT recompilation for unique tail-batch sizes.
-                raw_images, rotation_matrices, translations, ctf_params, noise_variance = (
-                    _pad_heterogeneity_kernel_batch(
-                        raw_images,
-                        rotation_matrices,
-                        translations,
-                        ctf_params,
-                        noise_variance,
-                        target_batch_size=batch_size,
-                    )
-                )
-                images = _process_images_half_fast(raw_images, experiment_dataset)
-            else:
-                images = experiment_dataset.process_images_half(raw_images)
-                images, rotation_matrices, translations, ctf_params, noise_variance = _pad_heterogeneity_kernel_batch(
-                    images,
-                    rotation_matrices,
-                    translations,
-                    ctf_params,
-                    noise_variance,
-                    target_batch_size=batch_size,
-                )
-            Ft_y_acc, Ft_ctf_acc = _heterogeneity_kernel_batch_from_fft(
-                config,
-                images,
-                ctf_params,
-                rotation_matrices,
-                translations,
-                noise_variance,
-                Ft_y=Ft_y_acc,
-                Ft_ctf=Ft_ctf_acc,
-            )
-
-        rhs_all[bin_idx] = np.asarray(Ft_y_acc)
-        lhs_all[bin_idx] = np.asarray(Ft_ctf_acc)
-
-    # A slight improvement is an almost triangular kernel/ pyramid kernel
-    #    _
-    #  _| |_
-    # _|     |_
-    # or almost epachenikov
-    #
-    # heterogeneity_kernel
-    if heterogeneity_kernel == "parabola" or heterogeneity_kernel == "triangle":
-        distances = bins
-        h_grid = 2 * bins
-
-        np_to_use = np
-
-        def kernel_fn(dist):
-            if heterogeneity_kernel == "triangle":
-                return np_to_use.where(np_to_use.abs(dist) < 1, 1 - np_to_use.abs(dist), 0)
-            return np_to_use.where(np_to_use.abs(dist) < 1, 3 / 4 * (1 - dist**2), 0)
-
-        weight_matrix = np_to_use.zeros((n_bins, n_bins)).astype(np.float32)
-        weight_matrix[0, 0] = 1
-        for idx in range(1, n_bins):
-            weights = kernel_fn(np_to_use.sqrt(distances / h_grid[idx]))
-            weight_matrix[:, idx] = weights
-        # Matmul on CPU: (50,50) @ (50, half_vol) — fast and avoids GPU OOM.
-        # Downstream post_process_from_filter_v2 transfers each row individually.
-        rhs_all = np.asarray(weight_matrix.T.astype(rhs_all.real.dtype) @ rhs_all)
-        lhs_all = np.asarray(weight_matrix.T @ lhs_all)
-
-    elif heterogeneity_kernel == "square":
-        rhs_all = np.cumsum(rhs_all, axis=0)
-        lhs_all = np.cumsum(lhs_all, axis=0)
-    else:
-        raise NotImplementedError
-
-    kernel_type = "triangular" if disc_type == "linear_interp" else "square"
-    vol_upsample = upsampling_factor if upsampling_factor is not None else 1
-    estimate_bins = []
-    for idx in range(heterogeneity_bins.size):
-        estimate = relion_functions.post_process_from_filter_v2(
-            lhs_all[idx],
-            rhs_all[idx],
-            experiment_dataset.volume_shape,
-            vol_upsample,
-            tau=tau,
-            kernel=kernel_type,
-            use_spherical_mask=use_spherical_mask,
-            grid_correct=grid_correct,
-            gridding_correct="square",
-            kernel_width=1,
-            return_real_space=return_real_space,
-            input_half_volume=True,
-        )
-        estimate_bins.append(estimate.reshape(-1))
-
-    estimates = np.asarray(jnp.stack(estimate_bins, axis=0))
-    if return_lhs_rhs:
-        return estimates, np.asarray(lhs_all), np.asarray(rhs_all)
-
-    return estimates
-
-
-def even_less_naive_deconvolved_heterogeneity_scheme_relion_style(
-    experiment_dataset,
-    signal_variance,
-    latent_differences,
-    latent_precision,
-    lambda_grid=None,
-    batch_size=None,
-    tau=None,
-    compute_lhs_rhs=False,
-    grid_correct=True,
-    disc_type="linear_interp",
-    use_spherical_mask=True,
-    return_lhs_rhs=False,
-    upsampling_factor=None,
-    return_real_space=False,
-    use_fast_rfft=False,
-    sigma_ref=None,
-    lambda_batch_size=None,
-):
-    del signal_variance, compute_lhs_rhs
-    latent_differences = _coerce_1d_latent_differences(latent_differences)
-    latent_precision = _coerce_1d_latent_precision(latent_precision)
-    latent_differences = _expand_tilt_latent_array_to_images(
-        experiment_dataset, latent_differences, "latent_differences"
-    )
-    latent_precision = _expand_tilt_latent_array_to_images(experiment_dataset, latent_precision, "latent_precision")
-    lambda_grid, h_grid, sigma_ref = deconvolution_bandwidths_1d(
-        latent_precision, lambda_grid=lambda_grid, sigma_ref=sigma_ref
-    )
-    n_lambdas = lambda_grid.size
-
-    if upsampling_factor is not None:
-        upsampled_vol_shape = tuple(3 * [experiment_dataset.grid_size * upsampling_factor])
-    else:
-        upsampled_vol_shape = tuple((experiment_dataset.grid_size,) * 3)
-    half_volume_size = np.prod(volume_shape_to_half_volume_shape(upsampled_vol_shape))
-    rhs_all = np.zeros((n_lambdas, half_volume_size), dtype=experiment_dataset.dtype)
-    lhs_all = np.zeros((n_lambdas, half_volume_size), dtype=experiment_dataset.dtype_real)
-
-    if batch_size is None:
-        accum_gb = utils.get_size_in_gb(rhs_all) + utils.get_size_in_gb(lhs_all)
-        avail_gb = _effective_heterogeneity_memory_budget(max(1.0, utils.get_gpu_memory_total() - accum_gb))
-        batch_size = int(utils.get_image_batch_size(experiment_dataset.grid_size, avail_gb))
-    if lambda_batch_size is None:
-        lambda_batch_size = _auto_deconv_lambda_batch_size(
-            n_lambdas,
-            half_volume_size,
-            experiment_dataset.dtype,
-            experiment_dataset.dtype_real,
-        )
-    lambda_batch_size = int(max(1, min(n_lambdas, lambda_batch_size)))
-    if custom_cuda_requested() and jax.default_backend() == "gpu" and core.decide_order(disc_type) <= 1:
-        batch_size = _auto_deconv_per_image_batch_size(
-            batch_size,
-            lambda_batch_size,
-            half_volume_size,
-            experiment_dataset.dtype,
-            experiment_dataset.dtype_real,
-        )
-    logger.info("batch size in deconvolved heterogeneity kernel: %s", batch_size)
-    logger.info("lambda batch size in deconvolved heterogeneity kernel: %s", lambda_batch_size)
-    logger.info("deconvolution lambda_grid=%s sigma_ref=%s", lambda_grid, sigma_ref)
-
-    if upsampling_factor is not None:
-        config = ForwardModelConfig.from_dataset(
-            experiment_dataset, disc_type=disc_type, upsampling_factor=upsampling_factor
-        )
-    else:
-        config = ForwardModelConfig.from_dataset(experiment_dataset, disc_type=disc_type, upsampling_factor=2)
-
-    for lambda_start in range(0, n_lambdas, lambda_batch_size):
-        lambda_stop = min(lambda_start + lambda_batch_size, n_lambdas)
-        h_group = h_grid[lambda_start:lambda_stop]
-        n_lambda_group = h_group.size
-        Ft_y_acc = jnp.zeros((n_lambda_group, half_volume_size), dtype=experiment_dataset.dtype)
-        Ft_ctf_acc = jnp.zeros((n_lambda_group, half_volume_size), dtype=experiment_dataset.dtype_real)
-        raw_batches = experiment_dataset.iter_batches(
-            batch_size,
-            noise_model=experiment_dataset.noise,
-            noise_half=False,
-        )
-        for (
-            raw_images,
-            rotation_matrices,
-            translations,
-            ctf_params,
-            noise_variance,
-            _particle_indices,
-            image_indices,
-        ) in raw_batches:
-            image_indices = np.asarray(image_indices, dtype=np.int32)
-            image_weights = deconvolution_weights_1d_many(
-                latent_differences[image_indices],
-                latent_precision[image_indices],
-                h_group,
-            )
-            if use_fast_rfft:
-                current_batch_size = int(raw_images.shape[0])
-                image_weights = _pad_image_weight_matrix_for_fixed_batch(
-                    image_weights,
-                    current_batch_size=current_batch_size,
-                    target_batch_size=batch_size,
-                )
-                raw_images, rotation_matrices, translations, ctf_params, noise_variance = (
-                    _pad_heterogeneity_kernel_batch(
-                        raw_images,
-                        rotation_matrices,
-                        translations,
-                        ctf_params,
-                        noise_variance,
-                        target_batch_size=batch_size,
-                    )
-                )
-                images = _process_images_half_fast(raw_images, experiment_dataset)
-            else:
-                images = experiment_dataset.process_images_half(raw_images)
-                current_batch_size = int(images.shape[0])
-                image_weights = _pad_image_weight_matrix_for_fixed_batch(
-                    image_weights,
-                    current_batch_size=current_batch_size,
-                    target_batch_size=batch_size,
-                )
-                images, rotation_matrices, translations, ctf_params, noise_variance = _pad_heterogeneity_kernel_batch(
-                    images,
-                    rotation_matrices,
-                    translations,
-                    ctf_params,
-                    noise_variance,
-                    target_batch_size=batch_size,
-                )
-            Ft_y_acc, Ft_ctf_acc = _heterogeneity_weighted_kernel_batch_many_from_fft(
-                config,
-                images,
-                ctf_params,
-                rotation_matrices,
-                translations,
-                noise_variance,
-                image_weights,
-                Ft_y=Ft_y_acc,
-                Ft_ctf=Ft_ctf_acc,
-            )
-
-        rhs_all[lambda_start:lambda_stop] = np.asarray(Ft_y_acc)
-        lhs_all[lambda_start:lambda_stop] = np.asarray(Ft_ctf_acc)
-
-    kernel_type = "triangular" if disc_type == "linear_interp" else "square"
-    vol_upsample = upsampling_factor if upsampling_factor is not None else 1
-    estimate_bins = []
-    for idx in range(n_lambdas):
-        estimate = relion_functions.post_process_from_filter_v2(
-            lhs_all[idx],
-            rhs_all[idx],
-            experiment_dataset.volume_shape,
-            vol_upsample,
-            tau=tau,
-            kernel=kernel_type,
-            use_spherical_mask=use_spherical_mask,
-            grid_correct=grid_correct,
-            gridding_correct="square",
-            kernel_width=1,
-            return_real_space=return_real_space,
-            input_half_volume=True,
-        )
-        estimate_bins.append(estimate.reshape(-1))
-
-    estimates = np.asarray(jnp.stack(estimate_bins, axis=0))
-    if return_lhs_rhs:
-        return estimates, np.asarray(lhs_all), np.asarray(rhs_all)
-
-    return estimates
-
-
-def compute_lhs_rhs(
-    cryo, noise_variance, heterogeneity_distances, residual_threshold, batch_size=100, disc_type="linear_interp"
-):
-
-    good_indices = heterogeneity_distances <= residual_threshold
-    test_dataset = cryoem_dataset.subsample_cryoem_dataset(cryo, good_indices)
-    Ft_ctf, F_ty = relion_functions.relion_style_triangular_kernel(
-        test_dataset, noise_variance.astype(np.float32), batch_size, disc_type=disc_type, upsampling_factor=1
-    )
-    return Ft_ctf, F_ty
 
 
 def estimate_variance_and_discretization_params(
