+++ tests/unit/test_analyze_jax_fallback_batch.py
@@ -121,19 +121,19 @@ def test_gpu_memory_arg_in_command(cmd_name):
 
 
 def test_effective_heterogeneity_memory_budget_passthrough(monkeypatch):
-    from recovar.heterogeneity import adaptive_kernel_discretization as akd
+    from recovar.heterogeneity import kernel_regression_reconstruction as kernel_recon
 
-    monkeypatch.setattr(akd, "custom_cuda_requested", lambda: True)
-    assert akd._effective_heterogeneity_memory_budget(48.0) == 48.0
+    monkeypatch.setattr(kernel_recon, "custom_cuda_requested", lambda: True)
+    assert kernel_recon._effective_heterogeneity_memory_budget(48.0) == 48.0
 
 
 def test_effective_heterogeneity_memory_budget_scales_for_fallback(monkeypatch, caplog):
-    from recovar.heterogeneity import adaptive_kernel_discretization as akd
+    from recovar.heterogeneity import kernel_regression_reconstruction as kernel_recon
 
-    monkeypatch.setattr(akd, "custom_cuda_requested", lambda: False)
-    caplog.set_level(logging.INFO, logger=akd.logger.name)
+    monkeypatch.setattr(kernel_recon, "custom_cuda_requested", lambda: False)
+    caplog.set_level(logging.INFO, logger=kernel_recon.logger.name)
 
-    scaled = akd._effective_heterogeneity_memory_budget(48.0)
+    scaled = kernel_recon._effective_heterogeneity_memory_budget(48.0)
 
     assert scaled == pytest.approx(16.0)
     assert "scaling heterogeneity-kernel memory budget" in caplog.text
